# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Spades multiplayer game
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RoomInput {
  playerName: string;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  playing: 'playing',
  finished: 'finished',
} as const;

export interface Room {
  roomCode: string;
  status: RoomStatus;
  playerCount: number;
  createdAt: string;
}

export interface ApiError {
  error: string;
}