# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Spades multiplayer game
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from './roomStatus';

export interface Room {
  roomCode: string;
  status: RoomStatus;
  playerCount: number;
  createdAt: string;
}