# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Spades multiplayer game
 * OpenAPI spec version: 0.1.0
 */

export interface RoomInput {
  playerName: string;
}